Moving inner classes back to IndexTankClient to ensure backwards compatibility
