+++ src/main/java/com/flaptor/indextank/apiclient/IndexAlreadyExistsException.java
@@ -1,8 +1,10 @@
 package com.flaptor.indextank.apiclient;
 
+import com.flaptor.indextank.apiclient.IndexTankClient.HttpCodeException;
+
 
 public class IndexAlreadyExistsException extends Exception {
-    public IndexAlreadyExistsException(ApiClient.HttpCodeException source) {
+    public IndexAlreadyExistsException(HttpCodeException source) {
         super(source.getMessage());
     }
 }
