+++ src/main/java/com/flaptor/indextank/apiclient/IndexDoesNotExistException.java
@@ -1,8 +1,10 @@
 package com.flaptor.indextank.apiclient;
 
+import com.flaptor.indextank.apiclient.IndexTankClient.HttpCodeException;
+
 
 public class IndexDoesNotExistException extends Exception {
-    public IndexDoesNotExistException(ApiClient.HttpCodeException source) {
+    public IndexDoesNotExistException(HttpCodeException source) {
         super(source.getMessage());
     }
 }
