+++ src/main/java/com/flaptor/indextank/apiclient/InvalidSyntaxException.java
@@ -1,8 +1,10 @@
 package com.flaptor.indextank.apiclient;
 
+import com.flaptor.indextank.apiclient.IndexTankClient.HttpCodeException;
+
 
 public class InvalidSyntaxException extends Exception {
-    public InvalidSyntaxException(ApiClient.HttpCodeException source) {
+    public InvalidSyntaxException(HttpCodeException source) {
         super(source.getMessage());
     }
 }
