+++ src/main/java/com/flaptor/indextank/apiclient/MaximumIndexesExceededException.java
@@ -1,10 +1,11 @@
 package com.flaptor.indextank.apiclient;
 
+import com.flaptor.indextank.apiclient.IndexTankClient.HttpCodeException;
+
 
 public class MaximumIndexesExceededException extends Exception {
 
-    public MaximumIndexesExceededException(
-            ApiClient.HttpCodeException source) {
+    public MaximumIndexesExceededException(HttpCodeException source) {
         super(source.getMessage());
     }
 }
