+++ src/main/java/com/flaptor/indextank/apiclient/UnexpectedCodeException.java
@@ -1,10 +1,12 @@
 package com.flaptor.indextank.apiclient;
 
+import com.flaptor.indextank.apiclient.IndexTankClient.HttpCodeException;
+
 
 public class UnexpectedCodeException extends RuntimeException {
     public final int httpCode;
 
-    public UnexpectedCodeException(ApiClient.HttpCodeException source) {
+    public UnexpectedCodeException(HttpCodeException source) {
         super(source.getMessage());
         this.httpCode = source.getHttpCode();
     }
